improve: Move ts types to interfaces folder
