+++ assets/interfaces/common.ts
@@ -1,0 +1,12 @@
+export type TDatetime = string; // ISO8601 format
+
+export interface IFilterGroup {
+    field: string;
+    label: string;
+    nested?: {
+        parent: string;
+        field: string;
+        value: string;
+        include_planning?: boolean;
+    };
+}
