+++ assets/interfaces/index.ts
@@ -1,0 +1,6 @@
+export {TDatetime, IFilterGroup} from './common';
+export {ICompany} from './company';
+export {IClientConfig} from './config';
+export {INavigation} from './navigation';
+export {ITopicNotificationScheduleType, ITopicFolder, ITopic} from './topic';
+export {IUserDashboard, IUser} from './user';
